Move admin api slice to a separate file
